KAN-111 [REFACTOR] Architectural Improvement on AppPreferenceRepository
  - Made architectural improvement on AppPreferenceRepository to prevent potential technical debt caused by monolithic Repository implementation.

    * Decentralize Monolithic AppPreferenceRepository to achieve strict Separation of Concern via Interface Segregation.
    * Necessary Binding functions have been defined for Dependency Injection via Dagger 2.
