+++ core/data/src/main/java/com/doyoonkim/data/di/LocalModules.kt
@@ -1,12 +1,18 @@
 package com.doyoonkim.data.di
 
 import android.content.Context
-import com.doyoonkim.data.repository.local.AppPreferenceRepositoryImpl
+import com.doyoonkim.data.repository.local.AppDatabasePreferenceRepositoryImpl
+import com.doyoonkim.data.repository.local.AppSubscriptionPreferenceRepositoryImpl
+import com.doyoonkim.data.repository.local.AppTokenPreferenceRepositoryImpl
+import com.doyoonkim.data.repository.local.AppWidgetPreferenceRepositoryImpl
 import com.doyoonkim.model.di.ApplicationContext
 import com.doyoonkim.data.repository.local.LocalRepositoryImpl
 import com.doyoonkim.data.repository.local.LocalWidgetCacheRepositoryImpl
 import com.doyoonkim.data.room.LocalDatabase
-import com.doyoonkim.domain.interfaces.AppPreferenceRepository
+import com.doyoonkim.domain.interfaces.AppDatabasePreferenceRepository
+import com.doyoonkim.domain.interfaces.AppSubscriptionPreferenceRepository
+import com.doyoonkim.domain.interfaces.AppTokenPreferenceRepository
+import com.doyoonkim.domain.interfaces.AppWidgetPreferenceRepository
 import com.doyoonkim.domain.interfaces.BookmarkLocalRepository
 import com.doyoonkim.domain.interfaces.LocalWidgetCacheRepository
 import com.doyoonkim.domain.interfaces.NoticeLocalRepository
@@ -54,7 +60,25 @@ object RoomDatabaseModule {
 abstract class LocalPreferenceModule {
     @Binds
     @Singleton
-    abstract fun bindLocalAppPreferenceRepository(
-        impl: AppPreferenceRepositoryImpl
-    ): AppPreferenceRepository
+    abstract fun bindLocalAppDatabasePreferenceRepository(
+        impl: AppDatabasePreferenceRepositoryImpl
+    ): AppDatabasePreferenceRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindLocalAppTokenPreferenceRepository(
+        impl: AppTokenPreferenceRepositoryImpl
+    ): AppTokenPreferenceRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindLocalAppSubscriptionPreferenceRepository(
+        impl: AppSubscriptionPreferenceRepositoryImpl
+    ): AppSubscriptionPreferenceRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindLocalAppWidgetPreferenceRepository(
+        impl: AppWidgetPreferenceRepositoryImpl
+    ): AppWidgetPreferenceRepository
 }
