+++ core/domain/src/main/java/com/doyoonkim/domain/interfaces/AppWidgetPreferenceRepository.kt
@@ -1,0 +1,15 @@
+package com.doyoonkim.domain.interfaces
+
+import com.doyoonkim.model.WidgetCategoryPolicy
+
+/**
+ * @author kimdoyoon
+ * Created 3/25/26 at 10:40 PM
+ */
+interface AppWidgetPreferenceRepository {
+
+    fun getWidgetCategoryPolicy(): WidgetCategoryPolicy
+
+    fun updateWidgetCategoryPolicy(policy: WidgetCategoryPolicy)
+
+}
